Health now use float

+++ Assets/Scripts/BuildSystem/Obstacle.cs
@@ -18,7 +18,7 @@ namespace bts {
       bar.SetUp(health);
     }
 
-    public void TakeDamage(int amount) {
+    public void TakeDamage(float amount) {
       health.Damage(amount);
       if (IsDead) {
         GridBuildingSystem.Demolish(transform.position);
@@ -31,7 +31,7 @@ namespace bts {
       }
     }
 
-    public void Heal(int amount) {
+    public void Heal(float amount) {
       health.Heal(amount);
     }
